backend/requirements.txt
fastapi==0.118.0
"fastapi[standard]"
pandas
uvicorn
asyncpg
python-dotenv
tortoise-orm[asyncpg]
passlib
bcrypt